+++ .run/[TCT] All tests in tests_trj_dataclasses dir.run.xml
@@ -13,7 +13,7 @@
     <option name="_new_keywords" value="&quot;&quot;" />
     <option name="_new_parameters" value="&quot;&quot;" />
     <option name="_new_additionalArguments" value="&quot;&quot;" />
-    <option name="_new_target" value="&quot;$PROJECT_DIR$/tests/tests_trj_dataclasses/&quot;" />
+    <option name="_new_target" value="&quot;$PROJECT_DIR$/tests/tests_dataclasses&quot;" />
     <option name="_new_targetType" value="&quot;PATH&quot;" />
     <method v="2" />
   </configuration>
