+++ tests/tests_dataclasses/tests_abstract_trj_dataclasses/__init__.py
@@ -1,0 +1,1 @@
+# coding=utf-8
