+++ tests/tests_dataclasses/tests_abstract_trj_dataclasses/test_abstract_multifeature_dataclass.py
@@ -8,14 +8,12 @@ import pytest
 
 from trajectory_container_tools import (
     AbstractMultifeatureDataclass,
-    BaseTrajectoryDataclass,
-)
+    )
 from trajectory_container_tools.dataclasses.ros2_feature_dataclass import (
     RosStampedDataclass,
 )
 from trajectory_container_tools.dataclasses.ros2_primitive_dataclass import Header
-from trajectory_container_tools.utils.temporal_tools.timestamps import Timestamps
-from trajectory_container_tools.utils.typing import TrajectoryDataclass
+from trajectory_container_tools.temporal.timestamps import Timestamps
 
 
 @dataclass
